Extract style component from Regions
