+++ src/RegionsContainer.jsx
@@ -2,30 +2,15 @@ import React from 'react';
 
 import { useDispatch, useSelector } from 'react-redux';
 
-import styled from '@emotion/styled';
-
 import {
   selectRegion,
   loadRestaurants,
 } from './slice';
 
+import Button from './style/Button';
+import Row from './style/Row';
+
 import { get } from './utils';
-
-const Row = styled.ul({
-  display: 'flex',
-  listStyle: 'none',
-  padding: '1rem 1rem',
-  overflowY: 'hidden',
-  '&::-webkit-scrollbar': {
-    width: 0,
-  },
-});
-
-const Button = styled.button({
-  margin: '0 .5rem',
-  padding: '.5rem 1rem',
-  width: '7rem',
-});
 
 export default function RegionsContainer() {
   const dispatch = useDispatch();
