+++ src/style/Button.jsx
@@ -1,0 +1,9 @@
+import styled from '@emotion/styled';
+
+const Button = styled.button({
+  margin: '0 .5rem',
+  padding: '.5rem 1rem',
+  width: '7rem',
+});
+
+export default Button;
